+++ src/lib-tempo/CMakeLists.txt
@@ -7,8 +7,8 @@ set(TEMPO_LIB_TEMPO_HEADERS
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/time.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/Types.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/component/ComponentGridAi.hpp"
-	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/ComponentGridMotion.hpp"
-	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/ComponentGridPosition.hpp"
+	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/component/ComponentGridMotion.hpp"
+	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/component/ComponentGridPosition.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/component/ComponentHealth.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/component/ComponentRender.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/component/ComponentTransform.hpp"
@@ -28,7 +28,7 @@ set(TEMPO_LIB_TEMPO_HEADERS
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/network/server.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/system/SystemGridAi.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/system/SystemHealth.hpp"
-	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/SystemLevelManager.hpp"
+	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/system/SystemLevelManager.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/system/SystemRender.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/system/SystemTransform.hpp"
 )
@@ -39,8 +39,8 @@ set(TEMPO_LIB_TEMPO_SOURCE
 	"${CMAKE_CURRENT_LIST_DIR}/src/song.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/structures.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/time.cpp"
-	"${CMAKE_CURRENT_LIST_DIR}/src/entity/ComponentGridMotion.cpp"
-	"${CMAKE_CURRENT_LIST_DIR}/src/entity/ComponentGridPosition.cpp"
+	"${CMAKE_CURRENT_LIST_DIR}/src/component/ComponentGridMotion.cpp"
+	"${CMAKE_CURRENT_LIST_DIR}/src/component/ComponentGridPosition.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/component/ComponentHealth.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/component/ComponentRender.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/component/ComponentTransform.cpp"
@@ -60,7 +60,7 @@ set(TEMPO_LIB_TEMPO_SOURCE
 	"${CMAKE_CURRENT_LIST_DIR}/src/network/server.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/system/SystemGridAi.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/system/SystemHealth.cpp"
-	"${CMAKE_CURRENT_LIST_DIR}/src/entity/SystemLevelManager.cpp"
+	"${CMAKE_CURRENT_LIST_DIR}/src/system/SystemLevelManager.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/system/SystemRender.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/system/SystemTransform.cpp"
 )
