+++ src/lib-tempo/include/tempo/component/ComponentGridPosition.hpp
