+++ src/lib-tempo/include/tempo/component/ComponentTransform.hpp
@@ -4,8 +4,8 @@
 #include <anax/System.hpp>
 #include <anax/Component.hpp>
 
-#include <tempo/entity/ComponentGridMotion.hpp>
-#include <tempo/entity/ComponentGridPosition.hpp>
+#include <tempo/component/ComponentGridMotion.hpp>
+#include <tempo/component/ComponentGridPosition.hpp>
 
 #include <OgreMath.h>
 #include <OgreVector3.h>
