+++ src/lib-tempo/include/tempo/entity/EntityCreationClient.hpp
@@ -12,7 +12,7 @@
 #include <tempo/entity/ID.hpp>
 #include <tempo/entity/PlayerLocal.hpp>
 #include <tempo/entity/PlayerRemote.hpp>
-#include <tempo/entity/SystemLevelManager.hpp>
+#include <tempo/system/SystemLevelManager.hpp>
 
 namespace tempo
 {
