+++ src/lib-tempo/src/component/ComponentGridPosition.cpp
@@ -1,4 +1,4 @@
-#include <tempo/entity/ComponentGridPosition.hpp>
+#include <tempo/component/ComponentGridPosition.hpp>
 
 namespace tempo{
 	/////////////////////////////////////////////////////////////
